+++ addOns/llm/CHANGELOG.md
@@ -5,8 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 ### Added
-- Allow to perform API sequencing and alert review.
 - Basic stats
+- Integration points for other add-ons.
+- Support for logging all LLM comms to a sub-tab of the main Output tab.
 
 ### Fixed
 - Error logs to always include stack trace.